refactor: add utils & plugin folder
