+++ app/controllers/api/index.js
@@ -3,7 +3,9 @@ const product = require("./product");
 const application = require("./application");
 const cities = require("./cities");
 const category = require("./category");
-const transaction = require("./transaction")
+const transaction = require("./transaction");
+const notification = require("./notification");
+const upload = require("./upload");
 
 module.exports = {
   authentication,
@@ -11,5 +13,7 @@ module.exports = {
   application,
   cities,
   category,
-  transaction
+  transaction,
+  notification,
+  upload
 };
