+++ app/middlewares/index.js
@@ -1,5 +1,7 @@
 const checkCondition = require("./checkCondition");
+const authorization = require("./authorization");
 
 module.exports = {
     checkCondition,
+    authorization,
 };
