+++ app/plugin/checkPassword.js
@@ -1,0 +1,7 @@
+const bcrypt = require("bcryptjs");
+
+async function checkPassword(password, encryptedPassword) {
+	return await bcrypt.compareSync(password, encryptedPassword)
+};
+
+module.exports = checkPassword;
