+++ app/plugin/index.js
@@ -1,0 +1,9 @@
+const checkPassword = require("./checkPassword");
+const createToken = require("./createToken");
+const hashPassword = require("./hashPassword");
+
+module.exports = {
+  checkPassword,
+  createToken,
+  hashPassword
+}
